feat(backup): Rename BackupConfig model to more general  Backup
models/Backup and models/Media will be the only two model files
for the moment.
