+++ src/app/domain/backup/models/index.ts
@@ -1,2 +1,2 @@
-export * from '/app/domain/backup/models/BackupConfig'
+export * from '/app/domain/backup/models/Backup'
 export * from '/app/domain/backup/models/Media'
